refector: Map 컴포넌트 기능만 통합시키고, 스타일을 따로 할 수 있게 작업
