+++ online_invitaion/src/components/sections/maps/Map.tsx
@@ -1,8 +1,10 @@
-import classNames from 'classnames/bind'
-import Section from '@/components/shared/sections/Section'
-import styles from './Map.module.scss'
 import { useEffect, useRef } from 'react'
 import { Location } from '@models/wedding'
+import classNames from 'classnames/bind'
+import styles from './Map.module.scss'
+
+// 타입에 따른 컴포넌트
+import MapType001 from './type/MapType001'
 
 const cx = classNames.bind(styles)
 
@@ -12,7 +14,7 @@ declare global {
   }
 }
 
-function Map({ location }: { location: Location }) {
+function Map({ location, type }: { location: Location; type: string }) {
   const mapContainer = useRef(null)
 
   // script 태그를 비동기적으로 호출하기 위한 코드 autoload가 비동기 sync가 어긋남을 직접 설정하기 위해서 false
@@ -45,54 +47,20 @@ function Map({ location }: { location: Location }) {
       })
     }
   }, [])
-  return (
-    <Section
-      title={
-        <div className={cx('wrap-header')}>
-          <span className={cx('txt-title')}>오시는길</span>
-          <span className={cx('txt-subtitle')}>{location.name}</span>
-          <span className={cx('txt-subtitle')}>{location.address}</span>
-        </div>
-      }
-    >
-      <div className={cx('wrap-map')}>
-        <div className={cx('map')} ref={mapContainer}></div>
-        <a
-          className={cx('btn-find-way')}
-          href={location.link}
-          target="_blank"
-          rel="noreferrer"
-        >
-          길찾기
-        </a>
-      </div>
 
-      <div>
-        <WayToCome label="버스" list={location.waytocome.bus} />
-        <WayToCome label="지하철" list={location.waytocome.metro} />
-      </div>
-    </Section>
-  )
-}
-
-// 찾아오는 길 컴포넌트
-function WayToCome({
-  label,
-  list,
-}: {
-  label: React.ReactNode
-  list: string[]
-}) {
-  return (
-    <div className={cx('wrap-waytocome')}>
-      <div className={cx('txt-label')}>{label}</div>
-      <ul>
-        {list.map((waytocome, idx) => (
-          <li key={idx}>{waytocome}</li>
-        ))}
-      </ul>
-    </div>
-  )
+  // 타입에 따라서 보여지는 맵 컴포넌트가 다르게
+  switch (type) {
+    case 'Type001':
+      return (
+        <MapType001 location={location}>
+          <div className={cx('type001_map')} ref={mapContainer}></div>
+        </MapType001>
+      )
+      break
+    default:
+      return null
+      break
+  }
 }
 
 export default Map
