+++ online_invitaion/src/components/sections/maps/type/MapType001.module.scss
@@ -1,0 +1,44 @@
+.wrap-header{
+    display: flex;
+    flex-direction: column;
+    text-align: left;
+
+    .txt-title{
+        margin-bottom: 16px;
+    }
+
+    .txt-subtitle{
+        font-weight: 400;
+        font-size: 15px;
+    }
+}
+
+.wrap-map{
+    position: relative;
+    
+    .btn-find-way{
+        border: 1px solid var(--black);
+        padding: 8px 32px;
+        z-index: 2;
+        position: absolute;
+        bottom: 24px;
+        left: 50%;
+        transform: translateX(-50%);
+        background-color: #fff;
+        color: var(--black);
+        text-decoration: none;
+        border-radius: 32px;
+    }
+}
+
+.wrap-waytocome{
+    margin: 16px 0;
+    line-height: 1.3;
+
+    .txt-label{
+        font-size: 18px;
+        font-weight: bold;
+        margin-bottom: 10px;
+        
+    }
+}
